refactor(api): keep warehouse row types on the metrics databricks port like sibling domains
