+++ apps/backend/src/metrics/core/models/experiment-organization.model.ts
@@ -1,0 +1,4 @@
+export interface ExperimentOrganizationRow {
+  experimentId: string;
+  organizationId: string | null;
+}
